# pkg/broadcastwriter: use []byte to lower alloc