```
benchmark                    old ns/op     new ns/op     delta
BenchmarkBroadcastWriter     1178209       669955        -43.14%
 
benchmark                    old MB/s     new MB/s     speedup
BenchmarkBroadcastWriter     21.00        36.93        1.76x
 
benchmark                    old allocs     new allocs     delta
BenchmarkBroadcastWriter     1510           5              -99.67%
 
benchmark                    old bytes     new bytes     delta
BenchmarkBroadcastWriter     64724         162           -99.75%
```

This PR refactors `pkg/broadcastwriter` to lower memory allocations and improve performance. It accomplishes this by avoiding the time.Format call for each loop and by avoiding allocations needed for the []byte to string conversions in various places.

This helps improve things further with #9139. These problems were even more obvious after #10347 was merged.

This entire change is one single commit right now, but it'll require discussion on how to document the new code in `pkg/jsonlog`.