@unclejack 2015-03-30
This PR fully fixes the extreme memory allocations I was seeing with `docker run -it --rm busybox sh -c "while true; do echo -n =; done"`. After running such a test, I've only got less than 500 MB of memory allocated (and released in the end) for the entire daemon running other containers as well. This used to be about 80 GB before.