@LK4D4 2015-03-30
Actually now I don't see where this branch of code used apart from `docker logs`. This is still important fix for docker logs, but could you pls write benchmark without `stderr` and `stdout` streams?