@unclejack 2015-03-30
@LK4D4 
```
+
+func BenchmarkBroadcastWriterWithoutStdoutStderr(b *testing.B) {
+       writer := New()
+       setUpWriter := func() {
+               for i := 0; i < 100; i++ {
+                       writer.AddWriter(devNullCloser(0), "")
+               }
+       }
+       testLine := "Line that thinks that it is log line from docker"
+       var buf bytes.Buffer
+       for i := 0; i < 100; i++ {
+               buf.Write([]byte(testLine + "\n"))
+       }
+       // line without eol
+       buf.Write([]byte(testLine))
+       testText := buf.Bytes()
+       b.SetBytes(int64(5 * len(testText)))
+       b.ResetTimer()
+       for i := 0; i < b.N; i++ {
+               b.StopTimer()
+               setUpWriter()
+               b.StartTimer()
+
+               for j := 0; j < 5; j++ {
+                       if _, err := writer.Write(testText); err != nil {
+                               b.Fatal(err)
+                       }
+               }
+
+               b.StopTimer()
+               writer.Clean()
+               b.StartTimer()
+       }
+}
```

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkBroadcastWriter                        1192551       559454        -53.09%
BenchmarkBroadcastWriterWithoutStdoutStderr     136957        65471         -52.20%

benchmark                                       old MB/s     new MB/s     speedup
BenchmarkBroadcastWriter                        20.75        44.22        2.13x
BenchmarkBroadcastWriterWithoutStdoutStderr     180.64       377.87       2.09x

benchmark                                       old allocs     new allocs     delta
BenchmarkBroadcastWriter                        1505           5              -99.67%
BenchmarkBroadcastWriterWithoutStdoutStderr     505            5              -99.01%

benchmark                                       old bytes     new bytes     delta
BenchmarkBroadcastWriter                        64482         162           -99.75%
BenchmarkBroadcastWriterWithoutStdoutStderr     32480         160           -99.51%
```