@LK4D4 2015-03-30
@unclejack I think cool optimization would be to not use `buf` logic at all if there is no `no-""` streams. Seems like now it used always, even if you don't read logs.