@unclejack 2015-03-31
@LK4D4 I've updated the code.

The new results for the benchmarks are below:
```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkBroadcastWriter                        1192293       542693        -54.48%
BenchmarkBroadcastWriterWithoutStdoutStderr     140267        14191         -89.88%

benchmark                                       old MB/s     new MB/s     speedup
BenchmarkBroadcastWriter                        20.75        45.59        2.20x
BenchmarkBroadcastWriterWithoutStdoutStderr     176.38       1743.35      9.88x

benchmark                                       old allocs     new allocs     delta
BenchmarkBroadcastWriter                        1505           5              -99.67%
BenchmarkBroadcastWriterWithoutStdoutStderr     609            104            -82.92%

benchmark                                       old bytes     new bytes     delta
BenchmarkBroadcastWriter                        64482         162           -99.75%
BenchmarkBroadcastWriterWithoutStdoutStderr     33728         1248          -96.30%
```