@josiasmontag 2015-04-03
Any chance that we see this in 1.6.0?