@unclejack 2015-04-03
@josiasmontag This should be in the next major release (1.7.0). If there's a 1.6.1 and if this change doesn't cause any kind of problems, it could be included in 1.6.1.

The time to get stuff into 1.6.0 has passed. Only fixes for potential critical bugs would get added to the 1.6.0 RC branch now.